# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Keep this title stable because it controls generated import paths.
  title: Api
  version: 1.0.0
  description: QuizIt competitive aptitude practice API
servers:
  - url: /api
    description: QuizIt API
...
paths:
...
  /topics/{topic_id}/subtopics:
    get:
      operationId: listSubtopics
      tags: [catalog]
      summary: List subtopics in a topic
      parameters:
        - $ref: "#/components/parameters/TopicId"
      responses:
        "200":
          description: Subtopics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subtopic"
  /questions:
    get:
      operationId: listQuestions
      tags: [catalog]
      summary: Browse active questions
      parameters:
        - name: topic_id
          in: query
          schema:
            type: integer
        - name: subtopic_id
          in: query
          schema:
            type: integer
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [easy, medium, hard]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Questions without answer keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
  /questions/{question_id}:
    get:
      operationId: getQuestion
      tags: [catalog]
      summary: Get one active question
      parameters:
        - $ref: "#/components/parameters/QuestionId"
      responses:
        "200":
          description: Question without answer key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
...
  /analytics/me:
    get:
      operationId: getMyAnalytics
      tags: [analytics]
      summary: Get the signed-in user's summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Personal analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"
  /analytics/me/history:
    get:
      operationId: getMyHistory
      tags: [analytics]
      summary: Get recent quiz attempts
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: topic_id
          in: query
          schema:
            type: [integer, "null"]
        - name: mode
          in: query
          schema:
            type: [string, "null"]
      responses:
        "200":
          description: Recent attempts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttemptSummary"
  /analytics/me/topics:
    get:
      operationId: getMyTopicInsights
      tags: [analytics]
      summary: Get accuracy by topic
      security:
        - bearerAuth: []
      parameters:
        - name: days
          in: query
          schema:
            type: [integer, "null"]
            minimum: 1
            maximum: 365
      responses:
        "200":
          description: Topic insights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopicInsight"
  /analytics/me/trend:
    get:
      operationId: getMyProgressTrend
      tags: [analytics]
      summary: Get a daily accuracy trend for the progress chart
      security:
        - bearerAuth: []
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            minimum: 7
            maximum: 180
            default: 30
      responses:
        "200":
          description: Daily accuracy points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgressTrendPoint"
  /analytics/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [analytics]
      summary: Get the leaderboard (global, the signed-in user's college, or friends)
      security:
        - bearerAuth: []
      parameters:
        - name: scope
          in: query
          schema:
            type: string
            enum: [global, college, friends]
            default: global
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Ranked users. The signed-in user is always included, even outside the top N.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    TopicId:
      name: topic_id
      in: path
      required: true
      schema:
        type: integer
    QuestionId:
      name: question_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    Subtopic:
      type: object
      required: [id, topic_id, name, slug]
      properties:
        id: { type: integer }
        topic_id: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: [string, "null"] }
    Question:
      type: object
      required: [id, subtopic_id, text, options]
      properties:
        id: { type: string }
        subtopic_id: { type: integer }
        text: { type: string }
        text_html: { type: [string, "null"] }
        options: {}
        options_html: {}
        answer_letter: { type: [string, "null"] }
        explanation: { type: [string, "null"] }
        explanation_html: { type: [string, "null"] }
        difficulty: { type: [string, "null"] }
...
    AnalyticsSummary:
      type: object
      required:
        [
          total_points, total_correct, total_incorrect, accuracy, matches_played, rating,
          current_streak, max_streak, total_xp, league, due_for_review,
        ]
      properties:
        total_points: { type: integer }
        total_correct: { type: integer }
        total_incorrect: { type: integer }
        accuracy: { type: number }
        matches_played: { type: integer }
        rating: { type: number }
        rank: { type: [integer, "null"] }
        recommended_topic: { type: [string, "null"] }
        current_streak: { type: integer }
        max_streak: { type: integer }
        total_xp: { type: integer }
        league: { type: string }
        due_for_review: { type: integer }
    AttemptSummary:
      type: object
      required: [attempt_id, quiz_id, quiz_mode, score, total_correct, total_incorrect, accuracy, completed_at]
      properties:
        attempt_id: { type: integer }
        quiz_id: { type: integer }
        quiz_mode: { type: string }
        score: { type: integer }
        total_correct: { type: integer }
        total_incorrect: { type: integer }
        accuracy: { type: number }
        completed_at: { type: [string, "null"], format: date-time }
    TopicInsight:
      type: object
      required: [topic_id, topic_name, total_answered, correct, accuracy]
      properties:
        topic_id: { type: integer }
        topic_name: { type: string }
        total_answered: { type: integer }
        correct: { type: integer }
        accuracy: { type: number }
    ProgressTrendPoint:
      type: object
      required: [date, attempts, correct, incorrect, accuracy]
      properties:
        date: { type: string, format: date }
        attempts: { type: integer }
        correct: { type: integer }
        incorrect: { type: integer }
        accuracy: { type: number }
    LeaderboardEntry:
      type: object
      required: [rank, user_id, username, name, total_points, user_rating, league]
      properties:
        rank: { type: integer }
        user_id: { type: integer }
        username: { type: string }
        name: { type: string }
        college_name: { type: [string, "null"] }
        total_points: { type: integer }
        user_rating: { type: number }
        league: { type: string }
        is_me: { type: boolean, default: false }
...